starting some frontend refactor
