+++ python/test/unit/operators/test_blocksparse.py
@@ -36,7 +36,6 @@ def mask_tensor(x, mask, block, value=0):
 @pytest.mark.parametrize("TRANS_A", [False, True])
 @pytest.mark.parametrize("TRANS_B", [False, True])
 @pytest.mark.parametrize("BLOCK", [16, 32, 64])
-# TODO: float32 fails
 @pytest.mark.parametrize("DTYPE", [torch.float16])
 def test_matmul(MODE, TRANS_A, TRANS_B, BLOCK, DTYPE, Z=3, H=2, M=512, N=384, K=256):
     seed = 0
@@ -82,14 +81,17 @@ def test_matmul(MODE, TRANS_A, TRANS_B, BLOCK, DTYPE, Z=3, H=2, M=512, N=384, K=
     a_tri.retain_grad()
     b_tri.retain_grad()
     op = triton.ops.blocksparse.matmul(layout, BLOCK, MODE, trans_a=TRANS_A, trans_b=TRANS_B, device="cuda")
-    c_tri = triton.testing.catch_oor(lambda: op(a_tri, b_tri), pytest)
-    triton.testing.catch_oor(lambda: c_tri.backward(dc_tri), pytest)
-    da_tri = a_tri.grad
-    db_tri = b_tri.grad
-    # compare
-    torch.testing.assert_allclose(c_ref, c_tri)
-    torch.testing.assert_allclose(da_ref, da_tri)
-    torch.testing.assert_allclose(db_ref, db_tri)
+    try:
+      c_tri = op(a_tri, b_tri)
+      c_tri.backward(dc_tri)
+      da_tri = a_tri.grad
+      db_tri = b_tri.grad
+      # compare
+      torch.testing.assert_allclose(c_ref, c_tri)
+      torch.testing.assert_allclose(da_ref, da_tri)
+      torch.testing.assert_allclose(db_ref, db_tri)
+    except triton.OutOfResourcesError as e:
+      pytest.skip(str(e))
 
 
 configs = [
