+++ python/triton/__init__.py
@@ -10,19 +10,15 @@ import torch  # noqa: F401
 
 # submodules
 from . import impl
-from .utils import (
-    cdiv,
-    MockTensor,
-    next_power_of_2,
-    reinterpret,
-    TensorWrapper,
-)
 from .runtime import (
     autotune,
     Config,
     heuristics,
     JITFunction,
     KernelInterface,
+    reinterpret,
+    TensorWrapper,
+    MockTensor,
 )
 from .runtime.jit import jit
 from .compiler import compile, CompilationError
@@ -49,3 +45,24 @@ __all__ = [
     "TensorWrapper",
     "testing",
 ]
+
+
+# -------------------------------------
+# misc. utilities that  don't fit well 
+# into any specific module 
+# -------------------------------------
+
+def cdiv(x, y):
+    return (x + y - 1) // y
+
+
+def next_power_of_2(n):
+    """Return the smallest power of 2 greater than or equal to n"""
+    n -= 1
+    n |= n >> 1
+    n |= n >> 2
+    n |= n >> 4
+    n |= n >> 8
+    n |= n >> 16
+    n += 1
+    return n
