+++ python/triton/runtime/driver/__init__.py
@@ -1,0 +1,2 @@
+from .cuda import get_cuda_utils
+from .hip import get_hip_utils
